Add support for newcomer priority
Newcomers get priority for their first song, with a cooldown. Fixed a bug which would make the queue order wrong if a user played a temporary song and then a playlist. Also fixed some major bugs that broke the server involving the new limits that I didn't notice because I didn't test enough.