Fix various bugs
Fixed the bug with song inactivity counts not resetting. Fixed a bug that could cause negative list indices when deleting things from queues or playlists. Fixed highlighting for the case where the song is playing, and distinguished whether the song is actually playing or just queued up. Updated the HTML and CSS a bit. Undid the title change. Stored song times to try to cut down on the YouTube scraping for video duration, which can cause problems if done too much. Fixed bug where deleting the currently playing song would mess up the index. Fixed bug where adding a song to the playlist while the last song is playing will go straight to the beginning of the playlist instead.